% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.downloadAppImage-package.R
\docType{package}
\name{rk.downloadAppImage-package}
\alias{rk.downloadAppImage}
\alias{rk.downloadAppImage-package}
\title{\packageTitle{rk.downloadAppImage}}
\description{
\packageDescription{rk.downloadAppImage}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab rk.downloadAppImage\cr
Type: \tab Package\cr
Version: \tab 0.01-0\cr
Date: \tab 2024-07-26\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
}

}
\author{
\packageAuthor{rk.downloadAppImage}

Maintainer: \packageMaintainer{rk.downloadAppImage}
}
